Refactor location of utils-files functions
Move some utils-functions to different files in order to
avoid depency cycle.
